+++ app/Http/Controllers/CashController.php
@@ -13,12 +13,14 @@ class CashController extends Controller
     public function index()
     {
         if (!auth()->check()) {
-            return redirect()->route('firstpage');
+            return redirect()->route('auth.login');
         }
-        $user = Auth::user();
-        $exchangeId = $user->exchange_id;
-        $cashRecords = Cash::where('exchange_id', $exchangeId)->get();
-        return view('exchange.cash.list',compact('cashRecords'));
+        else{
+            $user = Auth::user();
+            $exchangeId = $user->exchange_id;
+            $cashRecords = Cash::where('exchange_id', $exchangeId)->get();
+            return view('exchange.cash.list',compact('cashRecords'));
+        }
     }
 
     /**
@@ -35,37 +37,37 @@ class CashController extends Controller
     public function store(Request $request)
 {
     if (!auth()->check()) {
-        return response()->json(['error' => 'User not authenticated.'], 401);
+        return redirect()->route('auth.login');
     }
-
-    $validatedData = $request->validate([
-        'reference_number' => 'nullable|string|max:255|unique:cashes,reference_number',
-        'customer_name' => 'nullable|string|max:255',
-        'cash_amount' => 'required|numeric',
-        'cash_type' => 'required|in:deposit,withdrawal,expense',
-        'bonus_amount' => 'nullable|numeric',
-        'payment_type' => 'nullable|string',
-        'remarks' => 'required|string|max:255',
-    ]);
-
-    try {
-        $user = Auth::user();
-        Cash::create([
-            'reference_number' => $validatedData['reference_number'] ?? null,
-            'customer_name' => $validatedData['customer_name'] ?? null,
-            'cash_amount' => $validatedData['cash_amount'],
-            'cash_type' => $validatedData['cash_type'],
-            'bonus_amount' => $validatedData['bonus_amount'] ?? null,
-            'payment_type' => $validatedData['payment_type'] ?? null,
-            'remarks' => $validatedData['remarks'],
-            'user_id' => $user->id,
-            'exchange_id' => $user->exchange_id,
+    else{
+        $validatedData = $request->validate([
+            'reference_number' => 'nullable|string|max:255|unique:cashes,reference_number',
+            'customer_name' => 'nullable|string|max:255',
+            'cash_amount' => 'required|numeric',
+            'cash_type' => 'required|in:deposit,withdrawal,expense',
+            'bonus_amount' => 'nullable|numeric',
+            'payment_type' => 'nullable|string',
+            'remarks' => 'required|string|max:255',
         ]);
+        try {
+            $user = Auth::user();
+            Cash::create([
+                'reference_number' => $validatedData['reference_number'] ?? null,
+                'customer_name' => $validatedData['customer_name'] ?? null,
+                'cash_amount' => $validatedData['cash_amount'],
+                'cash_type' => $validatedData['cash_type'],
+                'bonus_amount' => $validatedData['bonus_amount'] ?? null,
+                'payment_type' => $validatedData['payment_type'] ?? null,
+                'remarks' => $validatedData['remarks'],
+                'user_id' => $user->id,
+                'exchange_id' => $user->exchange_id,
+            ]);
 
-        return response()->json(['message' => 'Data saved successfully!'], 200);
-    } catch (\Exception $e) {
-        \Log::error('Error while saving cash transaction: ' . $e->getMessage());
-        return response()->json(['error' => $e->getMessage()], 500);
+            return response()->json(['message' => 'Data saved successfully!'], 200);
+        } catch (\Exception $e) {
+            \Log::error('Error while saving cash transaction: ' . $e->getMessage());
+            return response()->json(['error' => $e->getMessage()], 500);
+        }
     }
 }
     
@@ -99,13 +101,16 @@ class CashController extends Controller
      */
     public function destroy(Request $request)
     {
-        $cash = Cash::find($request->id);
-        
-        if ($cash) {
-            $cash->delete();
-            return response()->json(['success' => true, 'message' => 'Cash deleted successfully!']);
+        if (!auth()->check()) {
+            return redirect()->route('auth.login');
         }
-
-        return response()->json(['success' => false, 'message' => 'Cank not found.'], 404);
+        else{
+            $cash = Cash::find($request->id);
+            if ($cash) {
+                $cash->delete();
+                return response()->json(['success' => true, 'message' => 'Cash deleted successfully!']);
+            }
+            return response()->json(['success' => false, 'message' => 'Cank not found.'], 404);
+        }
     }
 }
